chore(get-endpoint): renamed filenames and general cleanup
