+++ lib/plugins/features/movies/controller.js
@@ -11,18 +11,26 @@ exports.create = async (payload) => {
   return new Movie({ id: movie.id }).fetch();
 };
 
-exports.read = async (query) => {
+exports.list = async (query) => {
 
   return new Movie().query((qb) => {
 
+    // filter exact year
     if (query.year) {
-      // filter exact year
       qb.where('release_year', '=', query.year);
-    } else if (query.from_year && query.to_year) {
-      // filter between 2 years
-      qb.where('release_year', '>=', query.from_year).andWhere('release_year', '<=', query.to_year);
-    } if (query.title) {
-      // fuzzy filter by title
+    }
+
+    // filter between 2 years
+    if (query.from_year) {
+      qb.where('release_year', '>=', query.from_year);
+    }
+
+    if (query.to_year) {
+      qb.where('release_year', '<=', query.to_year);
+    }
+
+    // fuzzy filter by title
+    if (query.title) {
       qb.where('name', 'ILIKE', `%${query.title}%`);
     }
