+++ lib/plugins/features/movies/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const Controller        = require('./controller');
-const MovieValidator    = require('../../../validators/movie');
-const MovieGetValidator = require('../../../validators/movieget');
+const Controller         = require('./controller');
+const MovieValidator     = require('../../../validators/movies/create');
+const MovieListValidator = require('../../../validators/movies/list');
 
 exports.register = (server, options, next) => {
 
@@ -25,10 +25,10 @@ exports.register = (server, options, next) => {
       path: '/movies',
       config: {
         handler: (request, reply) => {
-          reply(Controller.read(request.query));
+          reply(Controller.list(request.query));
         },
         validate: {
-          query: MovieGetValidator
+          query: MovieListValidator
         }
       }
     }
