+++ package.json
@@ -33,6 +33,7 @@
     "test:raw": "NODE_ENV=test mocha test --require test/setup.js --recursive --timeout 30000"
   },
   "dependencies": {
+    "@hapi/joi": "^16.1.1",
     "bookshelf": "^0.15.1",
     "hapi": "16.6.1",
     "hapi-bookshelf-serializer": "2.1.0",
